FE tests refactoring
