+++ frontend/app/src/androidTest/java/com/cpen321/usermanagement/features/findJobs/FindJobsTestBase.kt
@@ -1,7 +1,6 @@
 package com.cpen321.usermanagement
 
 import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
@@ -12,70 +11,32 @@ import androidx.compose.ui.test.performTextInput
 import androidx.compose.ui.test.performTouchInput
 import androidx.compose.ui.test.swipeUp
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import androidx.test.uiautomator.By
-import androidx.test.uiautomator.UiDevice
-import androidx.test.uiautomator.Until
-import dagger.hilt.android.testing.HiltAndroidRule
+import com.cpen321.usermanagement.utils.BaseTestSetup
+import com.cpen321.usermanagement.utils.TestAccountHelper
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.Before
-import org.junit.Rule
 import org.junit.runner.RunWith
-import java.util.Properties
 
 /**
  * Base class for Find Jobs feature tests.
- * Provides common setup and utilities for testing.
- *
+ * Extends BaseTestSetup for standard test configuration.
+ * 
  */
 @HiltAndroidTest
 @RunWith(AndroidJUnit4::class)
-abstract class FindJobsTestBase {
-
-    @get:Rule(order = 0)
-    val hiltRule = HiltAndroidRule(this)
-
-    @get:Rule(order = 1)
-    val composeTestRule = createAndroidComposeRule<MainActivity>()
-
-    protected lateinit var device: UiDevice
+abstract class FindJobsTestBase : BaseTestSetup() {
 
     companion object {
-        private var testCredentials: Pair<String, String>? = null
-
-        private fun loadTestCredentials(): Pair<String, String> {
-            if (testCredentials != null) return testCredentials!!
-
-            val properties = Properties()
-            val inputStream = this::class.java.classLoader?.getResourceAsStream("test.properties")
-            if (inputStream != null) {
-                properties.load(inputStream)
-                val email = properties.getProperty("MOVER_EMAIL") ?: throw IllegalStateException("MOVER_EMAIL not found in test.properties")
-                val password = properties.getProperty("MOVER_PASSWORD") ?: throw IllegalStateException("MOVER_PASSWORD not found in test.properties")
-                testCredentials = email to password
-                return testCredentials!!
-            } else {
-                throw IllegalStateException("test.properties file not found in androidTest/resources")
-            }
-        }
-
-        fun getTestEmail(): String = loadTestCredentials().first
-        fun getTestPassword(): String = loadTestCredentials().second
+        fun getTestEmail(): String = TestAccountHelper.getMoverEmail()
+        fun getTestPassword(): String = TestAccountHelper.getMoverPassword()
     }
 
     @Before
-    fun baseSetup() {
-        hiltRule.inject()
-        device = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation())
-
-        // Wait for the app to settle after injection
-        composeTestRule.waitForIdle()
-
-        // Grant notification permission automatically
-        grantNotificationPermission()
+    override fun baseSetup() {
+        super.baseSetup()
 
         // Always need to be signed in as a mover for find jobs use cases
-        // check if we already see bio screen and sign in if not
+        // Check if we already see bio screen and sign in if not
         val completeProfileNodes = composeTestRule
             .onAllNodesWithText("Complete Your Profile", useUnmergedTree = true)
             .fetchSemanticsNodes()
@@ -91,281 +52,35 @@ abstract class FindJobsTestBase {
     }
 
     /**
-     * Sets up the test account by either signing up or adding the account to the device.
+     * Sets up the mover test account by either signing up or adding the account to the device.
      * Call this when you want to ensure the test account exists (for sign-up tests).
+     * Uses the MOVER account credentials from test.properties.
      */
     fun setupTestAccount() {
-        composeTestRule.waitForIdle()
-
-        val isNotAuthenticated = composeTestRule
-            .onAllNodesWithText("Sign in with Google")
-            .fetchSemanticsNodes()
-            .isNotEmpty()
-
-        if (isNotAuthenticated) {
-            // Click the sign-up button
-            composeTestRule.onNodeWithText("Sign up with Google", useUnmergedTree = true)
-                .assertExists("Sign up button should exist")
-                .performClick()
-
-            // Wait for Google account picker dialog to appear
-            val accountPickerAppeared = device.wait(
-                Until.hasObject(By.pkg("com.google.android.gms")),
-                10_000
-            )
-
-            if (accountPickerAppeared) {
-                Thread.sleep(1000) // Wait for UI to settle
-
-                val testEmail = FindJobsTestBase.Companion.getTestEmail()
-
-                // Look for the test account email
-                val testAccountElement = device.findObject(By.text(testEmail))
-
-                if (testAccountElement != null) {
-                    // Account exists, click it
-                    testAccountElement.click()
-                    Thread.sleep(2000) // Wait for consent screen to load
-
-                    // Wait for and click "Agree and share" button
-                    val agreeButton = device.wait(
-                        Until.findObject(By.textContains("Agree and share")),
-                        5_000
-                    )
-
-                    if (agreeButton != null) {
-                        agreeButton.click()
-                        Thread.sleep(2000)
-                    }
-                } else {
-                    // Account doesn't exist, need to add it
-                    // Look for "Add another account" or "Use another account"
-                    val addAccountButton = device.findObject(By.textContains("another account"))
-                        ?: device.findObject(By.textContains("Add account"))
-
-                    if (addAccountButton != null) {
-                        addAccountButton.click()
-                        Thread.sleep(3000) // Wait for Google sign-in page
-
-                        // Enter email - look for "Email or phone" field
-                        device.wait(
-                            Until.findObject(By.textContains("Email or phone")),
-                            5_000
-                        )
-
-                        // Now find the EditText input field
-                        var emailField = device.findObject(By.res("identifierId"))
-                        if (emailField == null) {
-                            emailField = device.findObject(By.clazz("android.widget.EditText"))
-                        }
-                        if (emailField == null) {
-                            emailField = device.findObject(By.focused(true))
-                        }
-
-                        if (emailField != null) {
-                            emailField.click() // Ensure it's focused
-                            Thread.sleep(1000)
-                            emailField.setText(testEmail)
-                            Thread.sleep(5000)
-
-                            // Try to find and click Next button - look for button specifically
-                            var nextButton = device.findObject(
-                                By.clazz("android.widget.Button").text("NEXT")
-                            )
-                            if (nextButton == null) {
-                                nextButton = device.findObject(
-                                    By.desc("NEXT")
-                                )
-                            }
-
-                            if (nextButton != null) {
-                                nextButton.click()
-                            } else {
-                                device.pressEnter()
-                            }
-                            Thread.sleep(5000)
-
-                            // Enter password - find the password EditText field
-                            var passwordField = device.findObject(By.res("password"))
-                            if (passwordField == null) {
-                                passwordField = device.findObject(By.clazz("android.widget.EditText"))
-                            }
-                            if (passwordField == null) {
-                                passwordField = device.findObject(By.focused(true))
-                            }
-
-                            if (passwordField != null) {
-                                passwordField.click()
-                                Thread.sleep(1000)
-                                val password = getTestPassword().replace(" ", "%s") // Escape spaces
-                                device.executeShellCommand("input text $password")
-                                Thread.sleep(5000)
-
-                                // Try to find and click Next button for password
-                                var nextButton = device.findObject(
-                                    By.clazz("android.widget.Button").text("NEXT")
-                                )
-                                if (nextButton == null) {
-                                    nextButton = device.findObject(
-                                        By.desc("NEXT")
-                                    )
-                                }
-
-                                if (nextButton != null) {
-                                    nextButton.click()
-                                } else {
-                                    device.pressEnter()
-                                }
-                                Thread.sleep(5000)
-
-                                // Handle Google account setup screens
-                                // Look for "Skip" button (recovery phone/email setup)
-                                var skipButton = device.findObject(By.text("Skip"))
-                                if (skipButton == null) {
-                                    skipButton = device.findObject(By.textContains("Skip"))
-                                }
-                                if (skipButton == null) {
-                                    skipButton = device.findObject(By.desc("Skip"))
-                                }
-
-                                if (skipButton != null) {
-                                    // Scroll to make sure it's visible
-                                    try {
-                                        skipButton.click()
-                                        Thread.sleep(5000)
-                                    } catch (e: Exception) {
-                                        // Button might not be clickable, continue
-                                    }
-                                }
-
-                                // Look for "I agree" button (terms and conditions)
-                                var agreeButton = device.findObject(By.text("I agree"))
-                                if (agreeButton == null) {
-                                    agreeButton = device.findObject(By.textContains("I agree"))
-                                }
-                                if (agreeButton == null) {
-                                    agreeButton = device.findObject(By.desc("I agree"))
-                                }
-
-                                if (agreeButton != null) {
-                                    // Scroll to make sure it's visible
-                                    try {
-                                        agreeButton.click()
-                                        Thread.sleep(5000)
-                                    } catch (e: Exception) {
-                                        // Button might not be clickable, continue
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            // Wait for sign-in flow to complete
-            composeTestRule.waitForIdle()
-            Thread.sleep(5000)
-
-
-           checkSetRole()
-
-            // Check if "Complete Your Profile" popup appears and skip it if present
-            val completeProfileNodes = composeTestRule
-                .onAllNodesWithText("Complete Your Profile", useUnmergedTree = true)
-                .fetchSemanticsNodes()
-
-            if (completeProfileNodes.isNotEmpty()) {
-                composeTestRule
-                    .onNodeWithText("Skip", useUnmergedTree = true)
-                    .performClick()
-            }
-
-            composeTestRule.waitForIdle()
-        }
+        TestAccountHelper.setupTestAccount(
+            composeTestRule = composeTestRule,
+            device = device,
+            email = getTestEmail(),
+            password = getTestPassword(),
+            roleSelector = { TestAccountHelper.selectMoverRole(it) }
+        )
     }
 
-    protected fun checkSetRole()
-    {
-        // Wait for role selection
-        val chooseRoleNode = composeTestRule
-            .onAllNodesWithText("I'm a Mover", useUnmergedTree = true)
-            .fetchSemanticsNodes()
-
-        // Choose mover role
-        if (chooseRoleNode.isNotEmpty()){
-            composeTestRule.onNodeWithText("I'm a Mover", useUnmergedTree = true)
-                .assertExists("Role button should exist")
-                .performClick()
-        }
-
-        composeTestRule.waitForIdle()
-    }
-
-    protected fun grantNotificationPermission() {
-        device.wait(
-            Until.findObject(By.text("Allow")),
-            5000
-        )?.click()
-    }
-
+    /**
+     * Signs in with the mover test account from test.properties.
+     * Looks for the specific test account in the account picker.
+     */
     private fun signInAsMover() {
-
+        // First ensure the account exists
         setupTestAccount()
 
-        composeTestRule.waitForIdle()
-
-        val isNotAuthenticated = composeTestRule
-            .onAllNodesWithText("Sign in with Google")
-            .fetchSemanticsNodes()
-            .isNotEmpty()
-
-        if (isNotAuthenticated) {
-            // Click the sign-in button
-            composeTestRule.onNodeWithText("Sign in with Google", useUnmergedTree = true)
-                .assertExists("Sign in button should exist (FindJobsTestBase)")
-                .performClick()
-
-            // Wait for Google account picker dialog to appear
-            val accountPickerAppeared = device.wait(
-                Until.hasObject(By.pkg("com.google.android.gms")),
-                10_000
-            )
-
-            if (accountPickerAppeared) {
-                Thread.sleep(1000) // Wait for UI to settle
-
-                val testEmail = FindJobsTestBase.Companion.getTestEmail()
-
-                // Look for the test account email
-                val testAccountElement = device.findObject(By.text(testEmail))
-
-                if (testAccountElement != null) {
-                    // Found the test account, click it
-                    testAccountElement.click()
-                }
-            }
-
-            // Wait for sign-in flow to complete
-            composeTestRule.waitForIdle()
-            Thread.sleep(2000)
-
-            checkSetRole()
-
-            // Check if "Complete Your Profile" popup appears and skip it if present
-            val completeProfileNodes = composeTestRule
-                .onAllNodesWithText("Complete Your Profile", useUnmergedTree = true)
-                .fetchSemanticsNodes()
-
-            if (completeProfileNodes.isNotEmpty()) {
-                // Try to find and click "Skip" or "Later" button
-                composeTestRule
-                    .onNodeWithText("Skip", useUnmergedTree = true)
-                    .performClick()
-            }
-
-            composeTestRule.waitForIdle()
-        }
-        // If already authenticated, continue
+        // Then sign in
+        TestAccountHelper.signIn(
+            composeTestRule = composeTestRule,
+            device = device,
+            email = getTestEmail(),
+            roleSelector = { TestAccountHelper.selectMoverRole(it) }
+        )
     }
 
     /**
@@ -502,7 +217,7 @@ abstract class FindJobsTestBase {
     /**
      * Add a time slot to a specific day
      */
-    protected fun addTimeSlot(day: String, startTime: String, endTime: String) {
+    protected fun addAvailabilityTimeSlot(day: String, startTime: String, endTime: String) {
         // Scroll until the day is visible
         scrollToDay(day)
         composeTestRule.waitForIdle()
